requirements.txt
fastapi
uvicorn[standard]
python-dotenv
chromadb
sentence-transformers
pdfplumber
google-genai
